Make microfs_get_ino() work correctly with special inodes.
Special inodes (block devs, char devs, FIFOs and sockets) should never
have a non-zero i_offset-value, so the offset parameter given to
microfs_get_ino() is instead used for them which points to the offset
of the given inode.

However, if this value is used as-is and there is a directory where
the first file is a special file (e.g. /dev), then the offset for that
special file will be the same as the i_offset-value for its parent
directory. This means that the parent inode will be returned by
microfs_get_inode() when the special file in question is requested,
creating a nasty circular reference.

The problem is easily fixed by adding a byte to the offset value,
which is how cramfs solves the problem.