Mute a compiler warning and fix two others.
The warning in devtable.h is about the "m"-modifier used for the sscanf-call,
it is a GNU extension and not part of the C standard - but it is still very
useful.

The other two warnings are to be found in microfs_inode.c and are caused by
-Wmaybe-uninitialized. It seems that it is the call to microfs_find_block()
that is the problem for the compiler, and it is correct that the params
blk_data_offset and blk_data_length can be left uninitialized if an I/O error
occurs, but in that case they are never used by microfs_readpage(), so it
should not be a problem. It does however seem unnecessary to argue with the
compiler over this issue when it is fixable by assigning the variables given
to microfs_find_block() to 0 first, two integer assignments can hardly hurt
performance significantly either.