Fix an image corruption issue in microfsmki.
If someone managed to fill a single directory with so many dentries that
the size of their data (that is; the metadata for the actual files) can
not fit in 24 bits (without hitting the MICROFS_MAXFILES limit in the
process), and this theoretical someone tried to make a microfs image out
of it, the resulting image would be corrupt.

This fix simply adds a check to walk_directory() that will terminate the
program if this unusual (and unlikely) case would ever occur.