Add tools for benchmarking microfs performance.
The basic idea behind this first and sketchy attempt to test and compare
performance is that every benchmark should be reproducible as long as the
seed that generated the benchmark in question is known.

It is written to be as unbiased as is practically possible. Disclaimer: It
may be very biased in practice.