Improve performance for block sizes bigger than PAGE_CACHE_SIZE.
The inspiration comes from studying how squashfs reads data and handles
decompression. The patch allows microfs to inflate data directly from the
buffer heads to the page cache pages, whenever possible, which can improve
performance for block sizes bigger than PAGE_CACHE_SIZE. One of the biggest
bottlenecks for microfs in the past was all the unnecessary copying of data
between different buffers, so this change is a significant improvement.

One small change in the image format was necessary to accommodate the new
read implementation: file holes are no longer skipped altogether in the
image, instead they are compressed by zlib and treated like any other data.
This is not a huge deal though, one should avoid the smaller block sizes
for performance reasons and to have microfsmki skip file holes was mostly
useful with smaller block sizes (as the holes are often quite small).