Add source_ERR to boilerplate.sh.
The error handling provided by boilerplate.sh is not very good at dealing
with errors in scripts included using "source <scriptname>.sh". The parent
script will have the previously executed command stored in _prev_command
until the "source" command has finished executing, so any error in the
"sourced" script will print the wrong command along with the line number
from it self.

To get around this one can use "source_ERR $0" in the beginning of the
script that is to be included. It will simply set _prev_command to
"source <scriptname>.sh", making it clearer what is happening and where
to look for the error.