Add support for stress/load testing.
It is now possible to give the flag "-S" to test.sh in order to have it
run a stress and/or load test. The idea is to try to test that the lkm
works as it should even when the system is under heavy load and have little
RAM to work with.

The contrast is an ordinary test run which pretty much allows the lkm
to run under ideal circumstances. Hopefully there will not be too many
bugs hiding between the extremes.