Optimize the image format.
This change moves all block pointers so that they get their own dedicated
"image section" rather than being placed just before the compressed blocks
that they point to. This makes it possible to use the block pointer cache
much more efficiently and it also means that the same blocks does not have
to be read twice when accessing the compressed blocks. The only downside
is that an extra block pointer pointing to the beginning of the first
compressed block is needed for all files, but 4 extra bytes per file is
a fair tradeoff for the small performance gain.

This means that the image is organized as follows:

 * Superblock (possibly preceded by padding)
 * Inodes/Dentries
 * Block pointers
 * Compressed blocks