Fix block sharing for regular files and symbolic links.
Up until this point block sharing has been _very_ broken, for many reasons:

 - Callbacks for qsort() were incorrectly implemented.
 - find_duplicates() would only work in the most trivial cases.
 - microfsmki and microfscki did not update the image size appropriately.

This patch should address these issues and make block sharing work as
intended. It might be a good idea to update the test suite to allow it to
perform additional checks after the basic checks are done, for example
to check that the microfsmki log reports that duplicate files were found
if that is to be expected.