Add thread support to sysdrain.
This change should improve the ability of sysdrain to drain resources on
systems with multiple processors and/or cores. It also makes sysdrain work
with bigger slot sizes for memory allocations to reach the nominal memory
drain quicker, which is important in order to reach the ceiling fast enough
on systems with a lot of RAM available.